requirements.txt
psycopg2-binary
paramiko
python-nmap
asyncio
requests
python-dotenv
ddtrace